requirements.txt
python-dotenv==0.14.0
pyserial==3.4
requests==2.21.0